requirements-prod.txt
# Core FastAPI
fastapi==0.119.0
uvicorn==0.37.0
starlette==0.48.0
python-multipart==0.0.20
gunicorn==23.0.0

# Base de datos
SQLAlchemy==2.0.44
PyMySQL==1.1.2
greenlet==3.2.4

# Validación y configuración
pydantic==2.12.3
pydantic_core==2.41.4
python-dotenv==1.1.1
email-validator==2.3.0

# Reconocimiento facial
# DeepFace 0.0.53 requiere TensorFlow 2.15.x con Keras 2.15.x
# Versiones específicas para evitar conflictos con Keras 3.x
tensorflow==2.15.0
tf-keras==2.15.0
keras==2.15.0
deepface==0.0.53

# Procesamiento de imágenes (requerido por DeepFace y validarRostroRapido)
# OpenCV 4.8.x es compatible con numpy < 2.0 (requerido por TensorFlow 2.15)
opencv-python==4.8.1.78
pillow==12.0.0
numpy==1.26.4  # TensorFlow 2.15 requiere numpy < 2.0.0
scipy==1.16.2

# Utilidades
requests==2.32.5
urllib3==2.5.0
certifi==2025.10.5
charset-normalizer==3.4.4
idna==3.11
cryptography
gdown==5.2.0  # Para descargar modelos desde Google Drive

# HTTP y networking
httptools==0.7.1
websockets==15.0.1
watchfiles==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
sniffio==1.3.1
anyio==4.11.0

# Utilidades generales
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.2
six==1.17.0
joblib==1.5.2
tqdm==4.67.1
lz4==4.4.4

# Type hints
typing_extensions==4.15.0
annotated-types==0.7.0
typing-inspection==0.4.2
